Dev update
App.java and AppController.java are the main classes, and App.fxml is the main node (a BorderPane).
In the BorderPane there are a menu at the top and the module at center, which the user can load from the menu.
So far there are the skeletton of the modules: Logger and Dualities, although they are not completely finished.
The enum type file Module.java contains the list of all the available modules.
What's next?  Create class Duality, Positive, Negative and Symmetric, Space, Time, And later create graphical representations of the TOD, TA and STC.